requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6 